+++ pkg/upgrade/upgrade_system_backend_urls_2.14.go
@@ -1,0 +1,22 @@
+package upgrade
+
+import (
+	appsv1 "github.com/openshift/api/apps/v1"
+
+	"github.com/3scale/3scale-operator/pkg/reconcilers"
+)
+
+// DeploymentConfigPodTemplateLabelsMutator ensures pod template labels are reconciled
+func SystemBackendUrls(desired, existing *appsv1.DeploymentConfig) (bool, error) {
+	var changed bool
+
+	// Remove old env var
+	oldVarRemoved := reconcilers.DeploymentConfigEnvVarReconciler(desired, existing, "BACKEND_ROUTE")
+
+	// Add new env vars
+	newVar1Added := reconcilers.DeploymentConfigEnvVarReconciler(desired, existing, "BACKEND_URL")
+	newVar2Added := reconcilers.DeploymentConfigEnvVarReconciler(desired, existing, "BACKEND_PUBLIC_URL")
+	changed = oldVarRemoved || newVar1Added || newVar2Added
+
+	return changed, nil
+}